+++ gg-ai-bot.py
@@ -1,115 +1,102 @@
-# from ollama import Client
 from ollama._client import Client
-import gradio as gr
 import os
+import gradio as gr
 
-LEARNERS_NAME = "HOORAIN"
 ENGLISH = "English"
 MATH = "Math"
 RESEARCH = "Research"
-CONTEXT = ""
+DEFAULT_LEARNER = "Hoorain"
 
-def prompt():
-    if CONTEXT == MATH:
-        return f'''
-        Assume the role of a teacher for a 9 years old homeschool child named {LEARNERS_NAME} who is learning Math from Khan academy and Beast Academy.
-        Your task is to help her in problem solving and critical thinking skills. 
-        You should not provide direct answers but help her with questions to think about the answers.
-        You are allowed to provide if her answer is correct or wrong.
-        You are not allowed to answer questions related to english, science, geography, or any general knowledge topics except math.
-        Please ensure to add more follow up questions to help her explore the answer.
-        Your name is "Guddu Guide".
-        '''
-    else: 
-        return f'''
-        Assume the role of a teacher for a 9 years old homeschool child named {LEARNERS_NAME} who is learning English.
-        Your task is to correct her grammar mistakes and help her use better english vocabulary. 
-        Also summarise her mistakes and share corrections. Dont be verbose.
-        You are not allowed to answer questions related to science, math, geography, or any general knowledge topics except english.
-        Please ensure to add more follow up questions to continue the conversation in English.
-        Appreciate {LEARNERS_NAME} if her sentences are gramatically correct. Your name is "Guddu Guide"
-        '''
+class GudduGuide:
+    def __init__(self):
+        self.ollama_client = Client(host=os.environ.get('OLLAMA_HOST'))
 
-OLLAMA_HOST = os.environ.get('OLLAMA_HOST')
-ollama_client = Client(host=OLLAMA_HOST)
 
-def generate_response(message, history):
-    formatted_history = []
-    ASSISTANT = "assistant"
-    USER = "user"
+    def prompt(self, name, context):
+        def learners_name(name):
+                if name is None or name.strip() == "":
+                    return DEFAULT_LEARNER
+                else:
+                    return name
+        
+        if context == MATH:
+            return f'''
+            Assume the role of a teacher for a 9 years old homeschool child named {learners_name(name)} who is learning Math from Khan academy and Beast Academy.
+            Your task is to help her in problem solving and critical thinking skills. 
+            You should not provide direct answers but help her with questions to think about the answers.
+            You are allowed to provide if her answer is correct or wrong.
+            You are not allowed to answer questions related to english, science, geography, or any general knowledge topics except math.
+            Please ensure to add more follow up questions to help her explore the answer.
+            Your name is "Guddu Guide".
+            '''
+        else: 
+            return f'''
+            Assume the role of a teacher for a 9 years old homeschool child named {learners_name(name)} who is learning English.
+            Your task is to correct her grammar mistakes and help her use better english vocabulary. 
+            Also summarise her mistakes and share corrections. Dont be verbose.
+            You are not allowed to answer questions related to science, math, geography, or any general knowledge topics except english.
+            Please ensure to add more follow up questions to continue the conversation in English.
+            Appreciate {learners_name(name)} if her sentences are gramatically correct. Your name is "Guddu Guide"
+            '''
 
-    def add_context(role, content):
-        formatted_history.append({"role": role, "content":content})    
+    def generate_response(self, message, history, name, context):
+        formatted_history = []
+        ASSISTANT = "assistant"
+        USER = "user"
 
-    add_context(ASSISTANT, prompt())
-    # Consider only last 10 message context
-    if history and len(history) > 0:
-        for user, assistant in history[-10:]:
-            add_context(USER, user)
-            add_context(ASSISTANT, assistant)
-        # formatted_history.append({"role": "user", "content": user })
-        # formatted_history.append({"role": "assistant", "content":assistant})
+        def add_context(role, content):
+            formatted_history.append({"role": role, "content":content})
 
-    # if history is None or len(history) == 0:
-    #     add_context(ASSISTANT, prompt())
-    # elif len(history[-1]) == 2:
-    #     user, assistant = history[-1]
-    #     add_context(ASSISTANT, f"{prompt()} \n users message: {user} \n assistance response: {assistant}")
-    # else:
-    #     add_context(ASSISTANT, f"{prompt()} \n\n {history[-1]}")
+        add_context(ASSISTANT, self.prompt(name, context))
 
-    add_context(USER, message)
-    
-    try: 
-        response = ollama_client.chat(
-            model='llama3.1',
-            messages=formatted_history,
-            stream=True,
+        if history and len(history) > 0:
+            for user, assistant in history[-10:]:
+                add_context(USER, user)
+                add_context(ASSISTANT, assistant)
+
+        add_context(USER, message)
+        
+        try: 
+            response = self.ollama_client.chat(
+                model='llama3.1',
+                messages=formatted_history,
+                stream=True,
+            )
+
+            partial_message = ""
+            for chunk in response:
+                if chunk['message']['content'] is not None:
+                        partial_message = partial_message + chunk['message']['content']
+                        yield partial_message
+        except Exception as e:
+            raise gr.Error("Guddu guide might be sleeping 💤🛌 Ask daddy to shake-it-up 🐣!", duration=10)
+
+    def chatbot(self, name, state):
+        return gr.ChatInterface(
+            self.generate_response,
+            additional_inputs=[name, state],
+            chatbot=gr.Chatbot(
+                label="Guddu Guide",
+                height=500
+            ),
+            textbox=gr.Textbox(
+                placeholder="You can ask me anything", 
+                container=False, 
+                scale=7
+            ),
+            retry_btn=None,
+            undo_btn=None,
+            clear_btn=None
         )
 
-        partial_message = ""
-        for chunk in response:
-            if chunk['message']['content'] is not None:
-                    partial_message = partial_message + chunk['message']['content']
-                    yield partial_message
-    except Exception as e:
-        raise gr.Error("Guddu guide might be sleeping 💤🛌 Ask daddy to shake-it-up 🐣!", duration=10)
-
-def chatbot():
-    return gr.ChatInterface(
-        generate_response,
-        chatbot=gr.Chatbot(
-            label="Guddu Guide",
-            height=500
-        ),
-        textbox=gr.Textbox(
-            placeholder="You can ask me anything", 
-            container=False, 
-            scale=7
-        ),
-        retry_btn=None,
-        undo_btn=None,
-        clear_btn=None
-    )
-
-if __name__ == "__main__":
+def main():
     theme = gr.themes.Soft(
         primary_hue="pink",
         secondary_hue="rose",
         neutral_hue="sky",
         text_size="lg",
     )
-
-    def change_name(name):
-        global LEARNERS_NAME
-        if name is None or name.strip() == "":
-            LEARNERS_NAME = "Hoorain"
-        else:
-            LEARNERS_NAME = name
-    
-    def change_tab(name):
-        global CONTEXT
-        CONTEXT = name
+    ggai = GudduGuide()
 
     with gr.Blocks(theme=theme, fill_height=True) as ggbot:
         gr.Markdown(
@@ -118,23 +105,33 @@ if __name__ == "__main__":
           There is no end to education. It is not that your ead a book, pass an exam and finish with education. The whole of life, from the moment you are born to the moment you die, is a process of learning.
         """)
         name_textbox = gr.Textbox(placeholder="add your name if you are not Hoorain", label=f"Hi there...")
-        name_textbox.change(change_name, name_textbox)
+        
+        # Next increment for applying themes
+        # with gr.Row():
+        #     with gr.Column():
+        #         theme = gr.Dropdown(["Blue", "Math", "Research"], info="What can I help you with?", show_label=False)
+        #     with gr.Column():
+        #         name_textbox = gr.Textbox(placeholder="add your name if you are not Hoorain", label=f"Hi there...")
 
-        with gr.Tab(ENGLISH) as english:
-            _ = chatbot()
-        with gr.Tab(MATH) as math:
-            _ = chatbot()
-        with gr.Tab(RESEARCH) as research:
-            _ = chatbot()
-        with gr.Tab("Why?") as about:
-            gr.Markdown(
-            """
-            # English
+        with gr.Tabs(visible=True, selected=ENGLISH): 
+            with gr.Tab(ENGLISH, id=ENGLISH) as english:
+                state = gr.State(ENGLISH)
+                _ = ggai.chatbot(name_textbox, state)
+            with gr.Tab(MATH, id=MATH) as math:
+                state = gr.State(MATH)
+                _ = ggai.chatbot(name_textbox, state)
+            with gr.Tab(RESEARCH, id=RESEARCH) as research:
+                state = gr.State(RESEARCH)
+                _ = ggai.chatbot(name_textbox, state)
+            with gr.Tab("Why?", id="about") as about:
+                gr.Markdown(
+                """
+                # English
 
-            # Math
+                # Math
 
-            # Research
-            """)
+                # Research
+                """)
         gr.Image(
                 "gg-tiny.png", 
                 scale=1, 
@@ -143,14 +140,11 @@ if __name__ == "__main__":
                 container=False, 
                 show_fullscreen_button=False
             )
-        
-        english.select(lambda :change_tab(ENGLISH), None)
-        math.select(lambda :change_tab(MATH), None)
-        research.select(lambda :change_tab(RESEARCH), None)
-        about.select(lambda :change_tab(ENGLISH), None)
-        CONTEXT = ENGLISH
     try:
         ggbot.launch(show_error=True)
     except Exception as e:
         print("An error occurred:", str(e))
+
+if __name__ == "__main__":
+    main()
